Changing price from int to double.

+++ Scrape items to buy/Websites/Retail Sites/Amazon.cs
@@ -65,7 +65,7 @@ namespace Scrape_items_to_buy.Websites.Sites
             return string.Empty;
         }
 
-        public override int GetPrice(string htmlBody)
+        public override double GetPrice(string htmlBody)
         {
             //*[@id=\"fk-mainbody-id\"]/div/div[8]/div/div[3]/div/div/div[4]/div/div[2]/div[1]/div/div[1]/div/div[1]/div/span[1]
             //*[@id="fk-mainbody-id"]/div/div[8]/div/div[3]/div/div/div[4]/div/div[2]/div[1]/div/div[1]/div/div[1]/span[1]
@@ -90,7 +90,7 @@ namespace Scrape_items_to_buy.Websites.Sites
                 node = root.SelectSingleNode("//*[@id='olp_feature_div']/div/span/span/text()");
             }
 
-            var price = Convert.ToInt32(Convert.ToDecimal(node.InnerHtml.Replace("Rs.", "").Replace(",", "").Replace(" ", "")));
+            var price = Convert.ToInt32(Convert.ToDouble(node.InnerHtml.Replace("Rs.", "").Replace(",", "").Replace(" ", "")));
             return price;
         }
